// frontend/src/api/generated/models/reservationRead.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * SpotPass Backend API
 * Restaurant reservation and management system
 * OpenAPI spec version: 1.0.0
 */
import type { ReservationReadAcceptedAt } from './reservationReadAcceptedAt';
import type { ReservationReadCanceledAt } from './reservationReadCanceledAt';
import type { ReservationReadNote } from './reservationReadNote';
import type { ReservationReadRefusedAt } from './reservationReadRefusedAt';
import type { ReservationReadReservationTime } from './reservationReadReservationTime';
import type { ReservationSource } from './reservationSource';
import type { ReservationReadSpecialRequest } from './reservationReadSpecialRequest';
import type { ReservationStatus } from './reservationStatus';

/**
 * Schema for reading reservations
 */
export interface ReservationRead {
  accepted_at?: ReservationReadAcceptedAt;
  canceled_at?: ReservationReadCanceledAt;
  created_at: string;
  establishment_id: string;
  id: string;
  no_show: boolean;
  note?: ReservationReadNote;
  /** @minimum 1 */
  number_of_guests: number;
  reference: string;
  refused_at?: ReservationReadRefusedAt;
  reservation_date: string;
  reservation_time?: ReservationReadReservationTime;
  source: ReservationSource;
  special_request?: ReservationReadSpecialRequest;
  status: ReservationStatus;
}